remove BooleanFile and add related method into test class
